Significant refactoring for Wiktor and Justin; overall functionality should be consistent with the initial commit.
