+++ src/Tests/Microsoft.Diagnostics.Monitoring.Tool.FunctionalTests/TestTimeouts.cs
@@ -6,7 +6,7 @@ using System;
 
 namespace Microsoft.Diagnostics.Monitoring.Tool.FunctionalTests
 {
-    public static class TestTimeouts
+    internal static class TestTimeouts
     {
         /// <summary>
         /// Default timeout for HTTP API calls
@@ -17,14 +17,6 @@ namespace Microsoft.Diagnostics.Monitoring.Tool.FunctionalTests
         /// Default logs collection duration.
         /// </summary>
         public static readonly TimeSpan LogsDuration = TimeSpan.FromSeconds(10);
-
-        /// <summary>
-        /// Default timeout for dump collection.
-        /// </summary>
-        /// <remarks>
-        /// Dumps (especially full dumps) can be quite large and take a significant amount of time to transfer.
-        /// </remarks>
-        public static readonly TimeSpan DumpTimeout = TimeSpan.FromMinutes(3);
 
         /// <summary>
         /// Timeout for polling a long running operation to completion.
